Remove wrong background plotter
