+++ tests/OTAnalytics/plugin_prototypes/track_visualization/test_track_viz.py
@@ -4,7 +4,6 @@ from unittest.mock import Mock, patch
 import pytest
 from pandas import DataFrame
 
-from OTAnalytics.application.datastore import Datastore
 from OTAnalytics.application.state import ObservableProperty, TrackViewState
 from OTAnalytics.domain.event import Event
 from OTAnalytics.domain.filter import Filter, FilterBuilder, FilterElement
@@ -41,7 +40,6 @@ from OTAnalytics.plugin_prototypes.track_visualization.track_viz import (
     PandasDataFrameProvider,
     PandasTrackProvider,
     PlotterPrototype,
-    TrackBackgroundPlotter,
     TrackGeometryPlotter,
     TrackPlotter,
     TrackStartEndPointPlotter,
@@ -334,35 +332,6 @@ class TestColorPaletteProvider:
 
         hex_color_pattern = re.compile(r"^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$")
         return bool(hex_color_pattern.match(value))
-
-
-class TestBackgroundPlotter:
-    def test_plot(self) -> None:
-        track = Mock(spec=Track).return_value
-        track.id = TrackId("5")
-
-        tracks = [track]
-        expected_image = Mock()
-        datastore = Mock(spec=Datastore)
-        datastore.get_all_tracks.return_value = tracks
-        datastore.get_image_of_track.return_value = expected_image
-
-        background_plotter = TrackBackgroundPlotter(datastore)
-        result = background_plotter.plot()
-
-        datastore.get_all_tracks.assert_called_once()
-        datastore.get_image_of_track.assert_called_once_with(track.id)
-        assert result is not None
-        assert result == expected_image
-
-    def test_plot_empty_track_repository_returns_none(self) -> None:
-        mock_datastore = Mock(spec=Datastore)
-        mock_datastore.get_all_tracks.return_value = []
-        background_plotter = TrackBackgroundPlotter(mock_datastore)
-        result = background_plotter.plot()
-
-        mock_datastore.get_all_tracks.assert_called_once()
-        assert result is None
 
 
 class TestTrackGeometryPlotter:
